Move some codes
