+++ src/llmperf/common.py
@@ -1,9 +1,12 @@
-from typing import List, Callable
+from typing import Any, Dict, List, Callable
+from llmperf.launcher import RequestsLauncher
+from llmperf.launcher.wait_for_all import WaitForAllLauncher
+from llmperf.launcher.wait_for_any import WaitForAnyLauncher
 from llmperf.ray_clients.furiosa_client import FuriosaLLMClient
 from llmperf.ray_clients.openai_chat_completions_client import (
     OpenAIChatCompletionsClient,
 )
-from llmperf.ray_llm_client import LLMClient
+from llmperf.ray_clients import LLMClient
 
 
 SUPPORTED_APIS = ["openai", "furiosa"]
@@ -35,3 +38,30 @@ def construct_clients(
         )
 
     return clients
+
+
+def construct_launcher(
+    wait_for: str,
+    model: str,
+    clients: List[LLMClient],
+    additional_sampling_params: Dict[str, Any],
+) -> RequestsLauncher:
+    """Construct RequestsLauncher that will send requests with a specific pattern.
+
+    Args:
+        wait_for: The name of pattern. WaitForAll launcher
+        model: The name of the model to query.
+        clients: The list of LLMClients.
+        additional_sampling_params: Additional sampling parameters to send with the request.
+            For more information see the LLM APIs documentation for the completions
+
+    Returns:
+        The constructed RequesstLauncher
+
+    """
+    if wait_for == "all":
+        return WaitForAllLauncher(model, clients, additional_sampling_params)
+    elif wait_for == "any":
+        return WaitForAnyLauncher(model, clients, additional_sampling_params)
+    else:
+        raise ValueError(f"Wrong type for 'wait_for' option: {wait_for}")
