+++ src/llmperf/ray_clients/openai_chat_completions_client.py
@@ -6,7 +6,7 @@ from typing import Any, Dict
 import ray
 import requests
 
-from llmperf.ray_llm_client import LLMClient
+from llmperf.ray_clients import LLMClient
 from llmperf.models import RequestConfig
 from llmperf import common_metrics
